requirements.txt
streamlit
numpy
pandas
singlestoredb
SQLAlchemy
sqlalchemy-singlestoredb
python-dotenv
